Added support for `jsonify` result to `OK` response in libprocess.
Review: https://reviews.apache.org/r/41594/
